pyproject.toml
[tool.poetry]
name = "jobify"
version = "1.1.0"
description = "Resume improvement application using Claude AI"
authors = ["Randall Sewell <mindaika@garbanzo.monster>"]

[tool.poetry.dependencies]
python = "^3.11"
flask = "3.0.0"
flask-cors = "4.0.0"
anthropic = "0.42.0"
pypdf2 = "3.0.1"
markdown = ">=3.5.2"
python-dotenv = ">=1.0.0"
werkzeug = "3.0.1"
httpx = ">=0.28.1"
python-jose = ">=3.3.0"
requests = "^2.32.3"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
certifi = "^2025.4.26"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
black = ">=23.12.0"
flake8 = ">=6.1.0"
pycparser = "^2.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"